Really resolves issue 30 (now allows floating-point values for songlist submenu x and y.
git-svn-id: http://fofix.googlecode.com/svn/MFH-Mod/trunk@331 48f48816-a574-47fe-9a59-6914b3aef992

+++ src/Theme.py
@@ -322,8 +322,8 @@ Config.define("theme", "fail_selected_color",  str, DEFAULT_FAIL_SELECTED_COLOR)
 Config.define("theme", "crowd_loop_delay",  int, DEFAULT_CROWD_LOOP_DELAY)
 
 #MFH - for instrument / difficulty / practice section submenu positioning
-Config.define("theme", "song_select_submenu_x",  int, DEFAULT_SONG_SELECT_SUBMENU_X)
-Config.define("theme", "song_select_submenu_y",  int, DEFAULT_SONG_SELECT_SUBMENU_Y)
+Config.define("theme", "song_select_submenu_x",  float, DEFAULT_SONG_SELECT_SUBMENU_X)
+Config.define("theme", "song_select_submenu_y",  float, DEFAULT_SONG_SELECT_SUBMENU_Y)
 
 #RACER:
 Config.define("theme", "fail_bkg_x",       float, DEFAULT_FAIL_BKG_X)
